Refactor and extraction of classes to own files.
