+++ source/Calamari/Deployment/PackageRetention/Repositories/JournalJsonRepository.cs
@@ -5,9 +5,8 @@ using System.Linq;
 using System.Text;
 using Calamari.Common.Features.Processes.Semaphores;
 using Calamari.Common.Plumbing.FileSystem;
+using Calamari.Deployment.PackageRetention.Model;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
-using SharpCompress.Common;
 
 namespace Calamari.Deployment.PackageRetention.Repositories
 {
